requirements-cloud.txt
# Web framework
Flask==3.0.0
gunicorn==21.2.0

# Google Cloud
google-cloud-bigquery==3.14.0
google-cloud-secret-manager==2.16.4
pandas-gbq==0.19.1

# Data processing
pandas==2.1.3
numpy==1.26.2

# Database
pyodbc==5.0.1

# API requests
requests==2.31.0

# Configuration
PyYAML==6.0.1

# Utilities
inflect==7.0.0
unidecode==1.3.0
tqdm==4.66.1

# Service account
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Development
python-dotenv==1.0.0
